Modularize MVC @Configuration
