+++ OpERP/src/main/java/devopsdistilled/operp/client/context/MvcControllerContext.java
@@ -1,0 +1,37 @@
+package devopsdistilled.operp.client.context;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import devopsdistilled.operp.client.items.panes.controllers.CreateCategoryPaneController;
+import devopsdistilled.operp.client.items.panes.controllers.CreateItemPaneController;
+import devopsdistilled.operp.client.items.panes.controllers.EditItemPaneController;
+import devopsdistilled.operp.client.items.panes.controllers.ListItemPaneController;
+import devopsdistilled.operp.client.items.panes.controllers.impl.CreateCategoryPaneControllerImpl;
+import devopsdistilled.operp.client.items.panes.controllers.impl.CreateItemPaneControllerImpl;
+import devopsdistilled.operp.client.items.panes.controllers.impl.EditItemPaneControllerImpl;
+import devopsdistilled.operp.client.items.panes.controllers.impl.ListItemPaneControllerImpl;
+
+@Configuration
+public class MvcControllerContext {
+
+	@Bean
+	public CreateItemPaneController createItemPaneController() {
+		return new CreateItemPaneControllerImpl();
+	}
+
+	@Bean
+	public ListItemPaneController listItemPaneController() {
+		return new ListItemPaneControllerImpl();
+	}
+
+	@Bean
+	public EditItemPaneController editItemPaneController() {
+		return new EditItemPaneControllerImpl();
+	}
+
+	@Bean
+	public CreateCategoryPaneController createCategoryPaneController() {
+		return new CreateCategoryPaneControllerImpl();
+	}
+}
