+++ OpERP/src/main/java/devopsdistilled/operp/client/context/MvcModelContext.java
@@ -1,0 +1,37 @@
+package devopsdistilled.operp.client.context;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import devopsdistilled.operp.client.items.panes.models.CreateCategoryPaneModel;
+import devopsdistilled.operp.client.items.panes.models.CreateItemPaneModel;
+import devopsdistilled.operp.client.items.panes.models.EditItemPaneModel;
+import devopsdistilled.operp.client.items.panes.models.ListItemPaneModel;
+import devopsdistilled.operp.client.items.panes.models.impl.CreateCategoryPaneModelImpl;
+import devopsdistilled.operp.client.items.panes.models.impl.CreateItemPaneModelImpl;
+import devopsdistilled.operp.client.items.panes.models.impl.EditItemPaneModelImpl;
+import devopsdistilled.operp.client.items.panes.models.impl.ListItemPaneModelImpl;
+
+@Configuration
+public class MvcModelContext {
+
+	@Bean
+	public CreateItemPaneModel createItemPaneModel() {
+		return new CreateItemPaneModelImpl();
+	}
+
+	@Bean
+	public ListItemPaneModel listItemPaneModel() {
+		return new ListItemPaneModelImpl();
+	}
+
+	@Bean
+	public EditItemPaneModel editItemPaneModel() {
+		return new EditItemPaneModelImpl();
+	}
+
+	@Bean
+	public CreateCategoryPaneModel createCategoryPaneModel() {
+		return new CreateCategoryPaneModelImpl();
+	}
+}
