+++ OpERP/src/main/java/devopsdistilled/operp/client/context/MvcViewContext.java
@@ -1,0 +1,34 @@
+package devopsdistilled.operp.client.context;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import devopsdistilled.operp.client.items.panes.CreateCategoryPane;
+import devopsdistilled.operp.client.items.panes.CreateItemPane;
+import devopsdistilled.operp.client.items.panes.EditItemPane;
+import devopsdistilled.operp.client.items.panes.ListItemPane;
+
+@Configuration
+public class MvcViewContext {
+
+	@Bean
+	public CreateItemPane createItemPane() {
+		return new CreateItemPane();
+	}
+
+	@Bean
+	public ListItemPane listItemPane() {
+		return new ListItemPane();
+	}
+
+	@Bean
+	public EditItemPane editItemPane() {
+		return new EditItemPane();
+	}
+
+	@Bean
+	public CreateCategoryPane createCategoryPane() {
+		return new CreateCategoryPane();
+	}
+
+}
